// src/medio-de-pago/enum/medio-de-pago.enum.ts
export enum TipoMedioDePagoEnum {
  EFECTIVO = 'EFECTIVO',
  TARJETA_CREDITO = 'TARJETA_CREDITO',
  TARJETA_DEBITO = 'TARJETA_DEBITO',
  TRANSFERENCIA_BANCARIA = 'TRANSFERENCIA_BANCARIA',
  CHEQUE = 'CHEQUE',
  CUENTA_CORRIENTE = 'CUENTA_CORRIENTE',
  OTRO = 'OTRO',
}
export enum RedDePago {
  VISA = 'VISA',
  MASTERCARD = 'MASTERCARD',
  AMERICAN_EXPRESS = 'AMERICAN_EXPRESS',
  NARANJA = 'NARANJA',
  MERCADOPAGO = 'MERCADOPAGO',
}